requirements.txt
# Telegram Bot Framework
aiogram>=3.4.0

# Database
asyncpg>=0.29.0

# HTTP Client
aiohttp>=3.9.0

# LangChain Core
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0

# Document Loaders
pypdf>=4.0.0
docx2txt>=0.8
unstructured>=0.12.0

# Vector Store
qdrant-client>=1.8.0

# OpenAI
openai>=1.12.0

# Environment Variables
python-dotenv>=1.0.0
